Connect pages map work ongoing
