+++ app/ATM/components/AtmNavbar.tsx
@@ -1,6 +1,7 @@
 'use client'
 
 import Image from 'next/image'
+import NextLink from 'next/link'
 
 import {
   Box,
@@ -55,7 +56,8 @@ export default function Navbar() {
           />
         </Flex>
         <Flex flex={{ base: 1 }} justify={{ base: 'center', md: 'start' }}>
-          <Button as={'a'} href={''} variant={'ghost'} size={'sm'} _hover={{bg: '00'}}>
+         
+          <Button as={'a'} href={'/'} variant={'ghost'} size={'sm'} _hover={{bg: '00'}}>
             <Image 
                 src={MazeBankLogo}
                 alt="Maze Bank Logo"
@@ -64,6 +66,7 @@ export default function Navbar() {
             />
             MAZE BANK
           </Button>
+          
 
           <Flex display={{ base: 'none', md: 'flex' }} ml={10}>
             <DesktopNav />
@@ -251,7 +254,7 @@ interface NavItem {
 
 const NAV_ITEMS: Array<NavItem> = [
   {
-    label: 'ATM Locations',
+    label: '',
     href: '#',
   },
 ]
