+++ app/providers.tsx
@@ -5,7 +5,6 @@ import Navbar from './components/Navbar'
 
 export function Providers({ children }: { children: React.ReactNode }) {
   return <ChakraProvider>
-        <Navbar/>
         {children}
     </ChakraProvider>
 }
